+++ resources/js/shopify-buysdk/checkout.js
